requirements.txt
fake_useragent==1.4.0
pandas==2.1.4
python-dotenv==1.0.0
Requests==2.31.0
selenium==4.16.0
